travelportals/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.8
django-appconf==1.0.5
django-compressor==2.4.1
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-decouple==2.1
django-filter==21.1
django-password-reset==2.0
django-social-auth==0.7.28
djangorestframework==3.12.4
easy-thumbnails==2.7.2
httpie==2.6.0
httplib2==0.20.1
idna==3.3
Markdown==3.3.4
oauth2==1.9.0.post1
Pillow==8.4.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==2.4.7
PySocks==1.7.1
python-openid==2.2.5
python3-openid==3.2.0
pytz==2021.3
raven==6.10.0
rcssmin==1.0.6
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rjsmin==1.1.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
urllib3==1.26.7
